[cli] drop cold-start detection and centralize trace analysis types
